requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
pandas>=2.0
openpyxl>=3.0
WeasyPrint>=59.0
reportlab>=4.0
PyPDF2>=3.0
Pillow>=10.0
python-docx>=0.8
playwright>=1.40
gunicorn>=21.0
